Huge refactor of opcode class names, preprocessing constants
Previous versions had a single "move instruction" rule, and there was a lot of logic that determined what opcode type would be produced. Due to the prior naming of the OpCode classes it wasn't obvious that Opcode 0/6/A were already handled and bundled into the "Assemble Move Instruction" method.

This refactoring along with producing different parser context objects for each opcode makes it more clear what has been added already.

Additionally this commit contains code for pre-processing the program and replacing any "constant" variables with their constant value prior to actual assembly. This serves as a way to disambiguate between rules that take [REG + REG] and [REG + NUM] when the 2nd value is a variable. The issue here is that [REG + VAR] can actually map to 2 different opcodes depending on VAR's type. This leads ANTLR into parsing the instruction as one opcode, and then after type checking the VAR we realized we need a different opcode.

By propagating constant values into the program *before* parsing for assembly, this ambiguity is solved. After propagation, if the parser finds any variable names they are dynamic ones.