Lots of fixes in the disassmbler/assembler
Disassembler tests exposed a large number of formatting inconsistencies (mostly around spacing around + operators). It also exposed a few issues where the disassembly was actually correct, but the assembler was wrong (which is where I got the test data for the disassembler tests). There are fixes throughout the opcodes/assembler/disassembler to address issues found by the unit tests.

At this point 85% of the Assembler is covered by tests and 78% of the Disassembler, and all tests are passing.